Refactor server structure: Rename and reorganize files
- Renamed `cardController.ts` to `controller.ts` and moved it to the controllers directory.
- Renamed `cardData.ts` to `data.ts` and moved it to the controllers directory.
- Renamed `CardStore.ts` to `entities/CardStore.ts` and updated import paths.
- Moved and renamed `cardRoute.ts` to `routes/index.ts`.
- Updated all import references.
